moved AbstractPCMWorkflowRunConfiguration to org.palladiosimulator.analyzer.workflow.core
