+++ bundles/org.palladiosimulator.analyzer.workflow.core/src/main/java/org/palladiosimulator/analyzer/workflow/core/jobs/IIssueReceiver.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.analyzer.workflow.jobs;
+package org.palladiosimulator.analyzer.workflow.core.jobs;
 
 import java.util.List;
