+++ bundles/org.palladiosimulator.analyzer.workflow/src/org/palladiosimulator/analyzer/workflow/jobs/PCMWorkflowJobBuilder.java
@@ -3,7 +3,7 @@
  */
 package org.palladiosimulator.analyzer.workflow.jobs;
 
-import org.palladiosimulator.analyzer.workflow.configurations.AbstractPCMWorkflowRunConfiguration;
+import org.palladiosimulator.analyzer.workflow.core.configurations.AbstractPCMWorkflowRunConfiguration;
 
 import de.uka.ipd.sdq.workflow.jobs.IJob;
